[Refactor] Перенос API и констант
